LayoutTests/js/dom/Promise-resolve-lookup-then-expected.txt

Test Promise.resolve looks up "then" synchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS count is 1
PASS error.message is "then lookup"
promise rejected.
PASS successfullyParsed is true

TEST COMPLETE
